Version 2 of data visualization
This is an improved version of mapping via geopandas and the natural earth map files originally implemented by Tasos. You will need to download these files as per the first version and put them into a folder named "mapdata" that is located in the same place as the data file and program file. Make sure everything is named correctly as it is in the code. Upon running the code you may automatically download additional maps that may be needed.