Basic grid of Denmark + vis
Created a visualization and mostly empty data file to represent a grid of Denmark. This is the same way the .nc weather files do it so it should all match up. I tried to remove any squares that weren't actually in Denmark. There are a few left to the south of Jutland but that should be fine. If you want to not have a random color grid change "random_color()" on line 31 to 'white', or whatever you want. Include the quotes.